rename sagas/discovery to sagas/disco (to encourage more dancing)
