+++ tests/unit/disco/sagas/testIndex.js
@@ -1,0 +1,9 @@
+import rootSagas from 'disco/sagas';
+
+describe(__filename, () => {
+  it('should run all sagas without an error', () => {
+    expect(() => {
+      rootSagas().next();
+    }).not.toThrowError();
+  });
+});
